Clarify that `--ask` is meaningless `nh ... build` or `nh home build --dry`.
This could be considered a bug report instead of a PR. I was confused
for a while why nh was ignoring my `--ask`, but then realized I was
doing `nh home build` instead of `nh home switch`.

It seems like `--ask --dry` is valid for `nh darwin`, so it is allowed.